Refactor due to module name change.
